Rearrange `drawContent()` contents
 + Moved the few lines of code inside the newly added "else" block.
 + Minor: "isTrailMode()", "isRgbMode()" and "isFill()" func() will receive the arguments instead of "event.target".

Even if both modes are turned off, the "for" loop inside "drawContent()" keeps running.
As there are no modes selected, it runs and throws error. Hence, moved this code to "else" block.
The "drawContent()" is the main function calling other modes as sub functions.
Thus, passing the target from the main to sub functions as a function parameter. Modified all sub functions to receive parameters.